refactor secular and root
