+++ pygrt/C_extension/src/modal/secular.c
@@ -13,7 +13,6 @@
 #include <stdlib.h>
 
 #include "grt/dynamic/layer.h"
-// #include "grt/dynamic/propagate.h"
 
 #include "grt/common/RT_matrix.h"
 #include "grt/common/model.h"
@@ -447,256 +446,3 @@ void grt_secular_function(
 {
     grt_secular_function_potential(mod1d, cphase, secRaylType, iref, wtype, psec, NULL);
 }
-
-
-
-
-
-size_t grt_goldensection_search_root(
-    GRT_MODEL1D *mod1d, const real_t c1, const real_t c2, 
-    const int secRaylType, const size_t iref, const DISPER_TYPE wtype,
-    const real_t rtol, cplx_t *root_sec, real_t *root_c)
-{
-    real_t c[2] = {c1, c2};
-    real_t x[2];
-    x[0] = c[1] - GOLDEN_RATIO*(c[1] - c[0]);
-    x[1] = c[0] + GOLDEN_RATIO*(c[1] - c[0]);
-    cplx_t sec[2];
-    real_t secABS[2];
-
-    for(int i=0; i<2; ++i){
-        grt_secular_function(mod1d, x[i], secRaylType, iref, wtype, &sec[i]);
-        secABS[i] = fabs(sec[i]);
-    }
-    size_t is=0;
-
-    #if __DEBUG_SECULAR__ == 1
-    fprintf(stderr, "# f=%f, root_sec2=%.9e%+.9ej|%.9e%+.9ej, c1=%.16e, c2=%.16e, dc=%.16e, root_c=%.9e, %d\n", 
-            creal(mod1d->omega)/PI2, creal(sec[0]), cimag(sec[0]), creal(sec[1]), cimag(sec[1]), x[0], x[1], x[1]-x[0], 
-            x[0], is);
-    fflush(stderr);
-    #endif
-    
-    do{
-        if(secABS[1] > secABS[0]){
-            c[1] = x[1];
-            sec[1] = sec[0];
-            secABS[1] = secABS[0];
-            x[1] = x[0];
-            x[0] = c[1] - GOLDEN_RATIO*(c[1] - c[0]);
-            grt_secular_function(mod1d, x[0], secRaylType, iref, wtype, &sec[0]);
-            secABS[0] = fabs(sec[0]);
-        } else {
-            c[0] = x[0];
-            sec[0] = sec[1];
-            secABS[0] = secABS[1];
-            x[0] = x[1];
-            x[1] = c[0] + GOLDEN_RATIO*(c[1] - c[0]);
-            grt_secular_function(mod1d, x[1], secRaylType, iref, wtype, &sec[1]);
-            secABS[1] = fabs(sec[1]);
-        }
-
-        // 接近达到浮点数精度极限
-        if(c[1] - c[0] < 1e-15)  break;
-
-        ++is;
-        #if __DEBUG_SECULAR__ == 1
-        fprintf(stderr, "# f=%f, root_sec2=%.9e%+.9ej|%.9e%+.9ej, secABS=%.9e, c1=%.16e, c2=%.16e, dc=%.19e, root_c=%.9e, %d\n", 
-            creal(mod1d->omega)/PI2, creal(sec[0]), cimag(sec[0]), creal(sec[1]), cimag(sec[1]), secABS[0], x[0], x[1], x[1]-x[0], 
-            x[0], is);
-        fflush(stderr);
-        #endif
-        
-    } while(is < 800);
-
-    #if __DEBUG_SECULAR__ == 1
-    fprintf(stderr, "# f=%f, root_sec2=%.9e%+.9ej|%.9e%+.9ej, dc=%e, root_c=%.9e, %d\n", 
-            creal(mod1d->omega)/PI2, creal(sec[0]), cimag(sec[0]), creal(sec[1]), cimag(sec[1]), x[1]-x[0], 
-            x[0], is);
-    fflush(stderr);
-    #endif
-
-    // 是否为零点
-    if(secABS[0] < rtol){
-        *root_sec = sec[0];
-        *root_c = x[0];
-        #if __DEBUG_SECULAR__ == 1
-        fprintf(stderr, "#      f=%f, root_sec=%.9e%+.9ej, root_c=%.9e, %d\n", creal(mod1d->omega)/PI2, creal(sec[0]), cimag(sec[0]), x[0], is);
-        fflush(stderr);
-        #endif
-    } else {
-        *root_sec = -999.0;
-        *root_c = -1.0;
-    }
-
-    return is;
-}
-
-
-
-// /** 
-//  * （仅备份）二分区间法确定久期函数零点 
-//  * 
-//  * @param[in]      mod1d        模型结构体指针
-//  * @param[in]      w            圆频率
-//  * @param[in]      kmid         区间中点波数
-//  * @param[in]      dk0          对称区间长度一半
-//  * @param[in]      secmid       区间中点久期函数值
-//  * @param[in]      secmidABS    区间中点久期函数绝对值
-//  * @param[in]      secRaylType  Rayl久期函数类型
-//  * @param[in]      iref         久期函数层位
-//  * @param[in]      isRayl       Rayleigh or Love
-//  * @param[out]     root_sec     零点的久期函数值
-//  * @param[out]     root_k       零点的波数
-//  * 
-//  */
-// size_t grt_bisearch_root(
-//     GRT_MODEL1D *mod1d, const real_t kmid, const real_t dk0,
-//     const cplx_t secmid, const real_t secmidABS, const int secRaylType, const size_t iref, const bool isRayl,
-//     const real_t rtol, cplx_t *root_sec, real_t *root_k)
-// {
-//     real_t k[3], dk;
-//     cplx_t sec[3];
-//     real_t secABS[3];
-//     bool recompute_sec[3] = {false, false, false};
-//     // MYINT dot3stat = 0;
-//     int inv_stats=GRT_INVERSE_SUCCESS;
-
-//     dk = dk0;
-//     k[1] = kmid;
-//     sec[1] = secmid;
-//     secABS[1] = secmidABS;
-//     dk *= 0.5;
-//     for(int i=0; i<=2; i+=2){
-//         k[i] = k[1] + (i-1)*dk;
-//         recompute_sec[i] = true;
-//     }
-
-//     size_t is=0;
-//     // char stats_str[2000]={'\0'};
-//     do{
-//         // 重新计算三点值
-//         for(int i=0; i<3; ++i){
-//             if(!recompute_sec[i])  continue;
-
-//             grt_secular_function(mod1d, w, k[i], secRaylType, iref, isRayl, &sec[i], &inv_stats);
-//             secABS[i] = fabs(sec[i]);
-//             recompute_sec[i] = false;
-//         }
-
-
-//         // 是否为零点，设计一个特别小的tol（tol < zero_tol）用于提前退出循环
-//         // 由于浮点数计算误差，零点的幅值不会无限逼近0，此处的tol是理想化误差，某些情况下的零点够不着
-//         if(secABS[1] < GRT_MIN(1e-8, GRT_ZEROROOT_TOL))  break;
-        
-//         // 判断修改方向，定义新的三点
-//         // 中间点为极小值点
-//         if(secABS[0] >= secABS[1] && secABS[1] <= secABS[2]){
-//             dk *= 0.5;
-//             // 更新两边
-//             for(int i=0; i<=2; i+=2){
-//                 k[i] = k[1] + (i-1)*dk;
-//                 recompute_sec[i] = true;
-//             }
-//             // dot3stat = 0;
-//             ++is;
-//         }
-//         // 左侧为极小值点
-//         else if(secABS[0] < secABS[1] && secABS[1] < secABS[2]){ 
-//             for(int i=2; i>0; --i){
-//                 k[i] = k[i-1];
-//                 sec[i] = sec[i-1];
-//                 secABS[i] = secABS[i-1];
-//             }
-//             k[0] = k[1] - dk;
-//             recompute_sec[0] = true;
-//             // dot3stat = 1;
-//         }
-//         // 右侧为极小值点
-//         else if(secABS[0] > secABS[1] && secABS[1] > secABS[2]){ 
-//             for(int i=0; i<2; ++i){
-//                 k[i] = k[i+1];
-//                 sec[i] = sec[i+1];
-//                 secABS[i] = secABS[i+1];
-//             }
-//             k[2] = k[1] + dk;
-//             recompute_sec[2] = true;
-//             // dot3stat = 2;
-//         }
-//         else {
-//             break;
-//         }
-
-//         // sprintf(stats_str, "%s %d", stats_str, dot3stat);
-        
-//     } while(is < 50);
-
-//     #if __DEBUG_SECULAR__ == 1
-//     fprintf(stderr, "# f=%f, root_sec3=%e%+ej|%e%+ej|%e%+ej, dk=%e, root_c=%f, %d\n", 
-//             w/PI2, creal(sec[0]), cimag(sec[0]), creal(sec[1]), cimag(sec[1]), creal(sec[2]), cimag(sec[2]), k[2]-k[0],
-//             w/k[1], is);
-//     fflush(stderr);
-//     #endif
-
-//     // 是否为零点
-//     if(secABS[1] < GRT_ZEROROOT_TOL){
-//         *root_sec = sec[1];
-//         *root_k = k[1];
-//         #if __DEBUG_SECULAR__ == 1
-//         fprintf(stderr, "#      f=%f, root_sec=%e%+ej, root_c=%f, %d\n", w/PI2, creal(sec[1]), cimag(sec[1]), w/k[1], is);
-//         fflush(stderr);
-//         #endif
-//     } else {
-//         *root_sec = -999.0;
-//         *root_k = -1.0;
-//     }
-
-//     return is;
-// }
-
-
-
-// /** 备份，固定间隔打印久期函数 */
-// int get_secular(int argc, char **argv){
-//     char *command = argv[0];
-
-//     char *s_modelpath = argv[1];
-//     real_t freq = atof(argv[2]);
-//     GRT_PYMODEL1D *pymod;
-//     // 读入模型文件（暂设置为不支持液体层）
-//     if((pymod = read_pymod_from_file(command, s_modelpath, -1, -1, true)) ==NULL){
-//         exit(EXIT_FAILURE);
-//     }
-//     // pymod1d -> mod1d
-//     GRT_MODEL1D *mod1d = init_mod1d(pymod->n);
-//     get_mod1d(pymod, mod1d);
-
-
-//     // Vs最大最小速度
-//     real_t vbmax = pymod->Vb[findMinMax_real_t(pymod->Vb, pymod->n, true)];
-//     real_t vbmin = pymod->Vb[findMinMax_real_t(pymod->Vb, pymod->n, false)];
-
-//     // print_mod1d(mod1d);
-
-//     cplx_t secRayl, secLove;
-//     MYINT secRaylType = 1;
-//     MYINT iref = 4;
-
-//     real_t omega, k;
-//     omega = PI2*freq;
-
-//     update_omega(mod1d, omega);
-//     // print_mod1d(mod1d);
-//     // getchar();
-
-//     MYINT inv_stats=GRT_INVERSE_SUCCESS;
-
-//     for(real_t c=vbmin*0.9; c<vbmax*1.1; c+=0.0001){
-//     // for(real_t c=3.4; c<3.9; c+=1e-5){
-//         k = omega/c;
-//         secular_function(mod1d, omega, k, secRaylType, iref, &secRayl, &secLove, &inv_stats);
-//         printf("%e %e %e %e %e \n", c, creal(secRayl), cimag(secRayl), creal(secLove), cimag(secLove));
-//     }
-
-//     exit(EXIT_SUCCESS);
-// }
